requirements.txt
pandas==2.1.3
yfinance==0.2.33
python-dotenv==1.0.0
azure-storage-blob==12.19.0